Fix: include file action use the same list of files as list files action
