+++ front/lib/api/assistant/jit_actions.ts
@@ -13,7 +13,6 @@ import type {
   ModelMessageTypeMultiActions,
   Result,
   RetrievalConfigurationType,
-  SupportedContentFragmentType,
   TablesQueryConfigurationType,
 } from "@dust-tt/types";
 import {
@@ -22,7 +21,6 @@ import {
   isAgentMessageType,
   isContentFragmentMessageTypeModel,
   isContentFragmentType,
-  isSupportedImageContentType,
   isUserMessageType,
   Ok,
   removeNulls,
@@ -36,11 +34,9 @@ import {
   DEFAULT_CONVERSATION_SEARCH_ACTION_DATA_DESCRIPTION,
   DEFAULT_CONVERSATION_SEARCH_ACTION_NAME,
 } from "@app/lib/api/assistant/actions/constants";
-import {
-  isConversationIncludableFileContentType,
-  makeConversationIncludeFileConfiguration,
-} from "@app/lib/api/assistant/actions/conversation/include_file";
+import { makeConversationIncludeFileConfiguration } from "@app/lib/api/assistant/actions/conversation/include_file";
 import { makeConversationListFilesAction } from "@app/lib/api/assistant/actions/conversation/list_files";
+import { listFiles } from "@app/lib/api/assistant/jit_utils";
 import {
   getTextContentFromMessage,
   getTextRepresentationFromMessages,
@@ -66,127 +62,6 @@ export async function isJITActionsEnabled(
   }
 
   return use;
-}
-
-function isQueryableContentType(
-  contentType: SupportedContentFragmentType
-): boolean {
-  if (isSupportedImageContentType(contentType)) {
-    return false;
-  }
-  // For now we only allow including text files.
-  switch (contentType) {
-    case "application/msword":
-    case "application/vnd.openxmlformats-officedocument.wordprocessingml.document":
-    case "application/pdf":
-    case "text/markdown":
-    case "text/plain":
-    case "dust-application/slack":
-    case "text/vnd.dust.attachment.slack.thread":
-    case "text/tab-separated-values":
-    case "text/tsv":
-      return false;
-
-    case "text/comma-separated-values":
-    case "text/csv":
-      return true;
-    default:
-      assertNever(contentType);
-  }
-}
-
-function isSearchableContentType(
-  contentType: SupportedContentFragmentType
-): boolean {
-  if (isSupportedImageContentType(contentType)) {
-    return false;
-  }
-  // For now we only allow including text files.
-  switch (contentType) {
-    case "application/msword":
-    case "application/vnd.openxmlformats-officedocument.wordprocessingml.document":
-    case "application/pdf":
-    case "text/markdown":
-    case "text/plain":
-    case "dust-application/slack":
-    case "text/vnd.dust.attachment.slack.thread":
-    case "text/tab-separated-values":
-    case "text/tsv":
-      return true;
-
-    case "text/comma-separated-values":
-    case "text/csv":
-      return false;
-    default:
-      assertNever(contentType);
-  }
-}
-
-function isListableContentType(
-  contentType: SupportedContentFragmentType
-): boolean {
-  // We allow listing all content-types that are not images. Note that
-  // `isSupportedPlainTextContentType` is not enough because it is limited to uploadable (as in from
-  // the conversation) content types which does not cover all non image content types that we
-  // support in the API such as `dust-application/slack`.
-  return !isSupportedImageContentType(contentType);
-}
-
-export function listFiles(
-  conversation: ConversationType
-): ConversationFileType[] {
-  const files: ConversationFileType[] = [];
-  for (const versions of conversation.content) {
-    const m = versions[versions.length - 1];
-
-    if (
-      isContentFragmentType(m) &&
-      isListableContentType(m.contentType) &&
-      m.contentFragmentVersion === "latest"
-    ) {
-      if (m.fileId) {
-        const canDoJIT = m.snippet !== null;
-        const isIncludable = isConversationIncludableFileContentType(
-          m.contentType
-        );
-        const isQueryable = canDoJIT && isQueryableContentType(m.contentType);
-        const isSearchable = canDoJIT && isSearchableContentType(m.contentType);
-
-        files.push({
-          fileId: m.fileId,
-          title: m.title,
-          contentType: m.contentType,
-          snippet: m.snippet,
-          isIncludable,
-          isQueryable,
-          isSearchable,
-        });
-      }
-    } else if (isAgentMessageType(m)) {
-      const generatedFiles = m.actions.flatMap((a) => a.getGeneratedFiles());
-
-      for (const f of generatedFiles) {
-        const canDoJIT = f.snippet != null;
-        const isIncludable = isConversationIncludableFileContentType(
-          f.contentType
-        );
-        const isQueryable = canDoJIT && isQueryableContentType(f.contentType);
-        const isSearchable = canDoJIT && isSearchableContentType(f.contentType);
-
-        files.push({
-          fileId: f.fileId,
-          contentType: f.contentType,
-          title: f.title,
-          snippet: f.snippet,
-          isIncludable,
-          isQueryable,
-          isSearchable,
-        });
-      }
-    }
-  }
-
-  return files;
 }
 
 async function getJITActions(
