+++ types/src/front/assistant/actions/conversation/list_files.ts
@@ -1,11 +1,15 @@
 import { BaseAction } from "../../../../front/assistant/actions/index";
-import { SupportedContentFragmentType } from "../../../../front/content_fragment";
+import {
+  ContentFragmentVersion,
+  SupportedContentFragmentType,
+} from "../../../../front/content_fragment";
 import { ModelId } from "../../../../shared/model_id";
 
 export type ConversationFileType = {
   fileId: string;
   title: string;
   contentType: SupportedContentFragmentType;
+  contentFragmentVersion: ContentFragmentVersion;
   snippet: string | null;
   isIncludable: boolean;
   isSearchable: boolean;
